refactor(nightshift): point the observability plugin at the extracted app
Observability no longer owns any Nightshift UI code.

- /nightshift route becomes a temporary redirect into /app/nightshift. This
  exists only so this PR is independently correct if it lands before the nav
  is repointed; the next PR deletes it, and no redirect ships to users.
- Agent-builder attachment registration moves into the nightshift plugin's
  start(). Keeping the call here would have meant a deep cross-plugin import
  into @kbn/nightshift-plugin/public/chat/..., i.e. an observability ->
  nightshift dependency that the next PR would immediately remove.
- Nav icon and the public re-export now come from observability_shared, so
  serverless_observability's import of NightshiftNavigationIcon from
  @kbn/observability-plugin/public keeps working unchanged.
- Global search provider import repointed at its parked location.
